Cloud_SQL_for_PostgreSQL/requirements.txt
langchain
langchain-community
langchain-core
langsmith
langgraph
langchain_google_genai
langchain_google_vertexai
langchain_google_community
langchain-google-cloud-sql-pg
pandas
unstructured
pyarrow